implemented bit.dev for component
