+++ .gitignore
@@ -1,5 +1,5 @@
 # See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
-
+.yarn
 # dependencies
 /node_modules
 /.pnp
