+++ src/components/confirm-button/index.js
@@ -1,5 +1,5 @@
 import React, { useState, useEffect, useCallback } from "react";
-import "../../style/custom.css";
+import "@vorteil/direktiv-ui.style/custom.css";
 import { IoCheckmarkSharp, IoCloseSharp } from "react-icons/io5"
 
 export function ConfirmButton(props) {
