+++ src/components/nav/index.js
@@ -1,6 +1,5 @@
 import React from 'react'
 import Logo from '@vorteil/direktiv-ui.img/direktiv.svg'
-import md5 from 'md5'
 import { Link, matchPath, useHistory, useLocation } from 'react-router-dom'
 
 import ArrowRightFill from 'react-bootstrap-icons/dist/icons/arrow-right-circle-fill'
@@ -23,19 +22,10 @@ export default function Navbar(props) {
         x.classList.toggle('active');
     }
 
-    // const [namespaces, setNamespaces] = useState([])
     const [acceptInput, setAcceptInput] = useState(false)
 
     const {fetch, namespace, setNamespace, namespaces, fetchNamespaces} = useContext(MainContext)
-    const {auth, name, email, logout} = props
-
-    let gravatarHash = ""
-    let gravatarURL = ""
-
-    if(auth) {
-        gravatarHash = md5(email)
-        gravatarURL = "https://www.gravatar.com/avatar/" + gravatarHash
-    }
+    const {footer} = props
     
 
     async function createNamespace(val) {
@@ -334,17 +324,7 @@ export default function Navbar(props) {
                     </li>
                 </ul>
             </div>
-            {auth ?
-        <div onClick={()=>logout()} id="nav-user-holder" style={{ paddingBottom: "40px" }}>
-            <img alt="user" style={{border: "solid 3px #0083B0", borderRadius: "50%", maxWidth: "48px"}} src={gravatarURL} />
-            <span style={{ fontSize: "10pt", display: "block" }}>
-                {name}
-            </span>
-        </div>
-            :   
-                            ""
-            }
-
+            {footer ? footer : "" }
         </div>
     )
 }
