+++ src/components/settings-page/index.js
@@ -1,12 +1,12 @@
 import React, { useCallback, useContext, useEffect, useState } from 'react'
-import Breadcrumbs from '../breadcrumbs'
-import TileTitle from '../tile-title'
-import MainContext from '../../context'
+import Breadcrumbs from '@vorteil/direktiv-ui.breadcrumbs'
+import TileTitle from '@vorteil/direktiv-ui.tile-title'
+import MainContext from '@vorteil/direktiv-ui.context'
 import { PlusCircle, XCircle } from 'react-bootstrap-icons'
 import { useHistory } from 'react-router'
 import { IoLockOpen, IoLogoDocker, IoTrash, IoWarningOutline } from 'react-icons/io5'
-import { sendNotification } from '../notifications'
-import { ConfirmButton, MiniConfirmButton } from '../confirm-button'
+import { sendNotification } from '@vorteil/direktiv-ui.notifications'
+import { ConfirmButton, MiniConfirmButton } from '@vorteil/direktiv-ui.confirm-button'
 
 
 function SettingsAction(props) {
