--- src/noresults.js
@@ -1,10 +1,0 @@
-// helper function to display no results in lists
-
-export default function NoResults() {
-
-    return(
-        <div style={{textAlign:"center"}}>
-            No results are found.
-        </div>
-    )
-}
